Refactor UI components.
