+++ web/app/src/main.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {createRoot} from "react-dom/client";
-import {App} from "./ui/app";
+import {App} from "./ui/App";
 
 import "../styles/chinese-checkers.less";
