+++ web/app/src/ui/board/GameBoardCell.tsx
@@ -1,0 +1,29 @@
+import clsx from "clsx";
+import {CellContent, inGreenTargetArea, inRedTargetArea} from "../../model/cell";
+import React from "react";
+import {Pawn} from "./Pawn";
+
+/**
+ * Game board cell.
+ */
+export function GameBoardCell({rowIndex, cellIndex, cell}: {
+	rowIndex: number;
+	cellIndex: number;
+	cell: CellContent;
+}) {
+	return (
+		<td
+			className={clsx({
+				"green-target": inGreenTargetArea(rowIndex, cellIndex),
+				"red-target": inRedTargetArea(rowIndex, cellIndex),
+			})}>
+			<button type={"button"}>
+				{ // Show a pawn if there is one.
+					cell != CellContent.Empty && (
+						<Pawn pawn={cell}/>
+					)
+				}
+			</button>
+		</td>
+	);
+}
