+++ web/app/src/ui/views/GameView.tsx
@@ -1,10 +1,9 @@
 import React, {useEffect} from "react";
 import {useParams} from "react-router-dom";
-import {usePageTitle} from "../layout";
-import {GameBoard} from "../components/games/game-board";
-import {Game, getCurrentPlayer} from "../../model/game";
-import {GamePlayer} from "../../model/game-player";
+import {usePageTitle} from "../Layout";
+import {GameBoard} from "../board/GameBoard";
 import {useFetchGame} from "../../api/games";
+import {PlayerTurn} from "../board/PlayerTurn";
 
 /**
  * Game view component.
@@ -27,17 +26,5 @@ export function GameView() {
 			<PlayerTurn game={fetchedGame.data}/>
 			<GameBoard board={fetchedGame.data.board}/>
 		</main>
-	);
-}
-
-export function PlayerTurn({game}: {
-	game: Game;
-}) {
-	const currentPlayer = getCurrentPlayer(game);
-	return (
-		<p className={"player-turn"}>
-			<strong
-				className={currentPlayer.gamePlayer == GamePlayer.Green ? "green" : "red"}>{currentPlayer.name}</strong> to play
-		</p>
 	);
 }
